# ordino/swagger.yml
swagger: '2.0'

# This is your document metadata
info:
  version: '2.1.1'
  title: Target Discovery Platform API

produces:
  - application/json

tags:
  - name: db
    description: database api
  - name: storage
    description: named set storage api
  - name: proxy
    description: proxy api

paths:
  /:
    get:
      x-swagger-router-controller: ordino
      operationId: sql.list_databases
      description: list all known databases
      tags:
        - db
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfDatabases
            type: array
            items:
              $ref: '#/definitions/database'
  /{database}:
    parameters:
      - $ref: '#/parameters/database'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.list_view
      description: list all views of a given database
      tags:
        - db
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfViews
            type: array
            items:
              $ref: '#/definitions/view'
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.get_data_api
      description: execute the view and return the data the arguments depend on the definition
      tags:
        - db
      parameters:
        - $ref: '#/parameters/_assignids'
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfRows
            type: array
            items:
              type: object
              additionalProperties: true
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}/filter:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.get_filtered_data
      description: see the base version with additional filter possiblities
      tags:
        - db
      parameters:
        - $ref: '#/parameters/_assignids'
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfRows
            type: array
            items:
              type: object
              additionalProperties: true
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}/score:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.get_score_data
      description: see the base version with additional filter possiblities and score mapping
      tags:
        - db
      parameters:
        - $ref: '#/parameters/_assignids'
        - name: target
          in: query
          type: string
          required: false
          description: target idtype to map the resulting scores to
      responses:
        '200':
          description: Successful response
          schema:
            title: ArrayOfRows
            type: array
            items:
              type: object
              additionalProperties: true
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}/count:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.get_count_data
      description: see the base version but returns just the number of results
      tags:
        - db
      responses:
        '200':
          description: Successful response
          schema:
            title: number of records
            type: integer
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}/desc:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.get_desc
      description: returns the description information of the given view
      tags:
        - db
      responses:
        '200':
          description: Successful response
          schema:
            title: column description
            type: object
            properties:
              idType:
                type: string
                description: the IDType of resulting rows
              columns:
                $ref: '#/definitions/columns'
        '404':
          $ref: '#/responses/notFound'
  /{database}/{view}/lookup:
    parameters:
      - $ref: '#/parameters/database'
      - $ref: '#/parameters/view'
    get:
      x-swagger-router-controller: ordino
      operationId: sql.lookup
      description: lookups up the given query term using this view
      tags:
        - db
      parameters:
        - $ref: '#/parameters/_assignids'
        - name: query
          in: query
          type: string
          required: true
          description: the search string
        - name: page
          in: query
          type: integer
          required: false
          default: 0
          description: result page
        - name: limit
          in: query
          type: integer
          required: false
          default: 30
          description: number of results per page
      responses:
        '200':
          description: Successful response
          schema:
            title: column description
            type: object
            properties:
              more:
                type: boolean
                description: whether more results exists
              items:
                type: object
                properties:
                  _id:
                    type: integer
                    description: the system wide unique id
                  id:
                    type: string
                    description: the id of the found entry
                  text:
                    type: string
                    description: label that matched the given query
        '404':
          $ref: '#/responses/notFound'

parameters:
  database:
    name: database
    description: the database key
    in: path
    required: true
    type: string
  view:
    name: view
    description: the view name
    in: path
    required: true
    type: string
  _assignids:
    name: _assignids
    in: query
    description: whether the unique ids should be automatically injected in the results
    required: false
    type: boolean
    default: false

definitions:
  database:
    type: string
  view:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      arguments:
        type: array
        items:
          - type: string
      query:
        type: string
      columns:
        $ref: '#/definitions/columns'
      idType:
        type: string
      filters:
        type: array
        items:
          - type: string
      queries:
        type: object
        additionalProperties: true
    required:
      - name
      - description
      - arguments
      - query
  columns:
    type: array
    description: list of columns
    items:
      - $ref: '#/definitions/column'
  column:
    type: object
    properties:
      column:
        type: string
      label:
        type: string
      type:
        type: string
        enum:
          - categorical
          - number
          - string
      min:
        type: number
      max:
        type: number
      categories:
        type: array
        items:
          - type: string
    required:
      - column
      - type
      - label
    additionalProperties: true

responses:
  error:
    description: Internal Server Error
    schema:
      title: Error
      type: string
  invalid:
    description: Invalid Request
    schema:
      title: Error
      type: string
  notFound:
    description: Not Found
    schema:
      title: Error
      type: string